Adding function that returns euclidian distance.
Adding a getter for the euclidian distance between nodes. Also added a  test that show that the function is doing its job propperly, since there are only two nodes in the test the euclidian distance between them is the middle of the nodes thusly there is no difference between them.